Add Media service and cleanup play command
