+++ src/commands/play/index.ts
@@ -1,9 +1,5 @@
-import { Command } from "../../models/command";
-import * as ytdl from "ytdl-core";
-import { SongInfo } from "../../models/song";
 import { TYPES } from "../../types";
 import { inject, injectable } from "inversify";
-import { PlayerService } from "../../services/player";
 import {
   ActionRowBuilder,
   ButtonBuilder,
@@ -15,12 +11,10 @@ import {
   SlashCommandBuilder,
   VoiceBasedChannel,
 } from "discord.js";
-import { Artist, SpotifyApi } from "@spotify/web-api-ts-sdk";
-import { Client as YouTubeIClient, VideoCompact } from "youtubei";
 import IOCContainer from "../../inversify.config";
-import { GuildQueueService } from "../../services/queue";
 import { formatSecondsToReadableTime } from "../../utils";
-import { ServerQueue } from "../../models/queue";
+import { GuildQueueService, MediaService, PlayerService } from "../../services";
+import { Command, ServerQueue, SongInfo } from "../../models";
 
 enum PlayCommandMessageComponentID {
   NEXT_SONG = "play.nextSongButton",
@@ -32,9 +26,7 @@ enum PlayCommandMessageComponentID {
 export default class PlayCommand implements Command {
   private queueService: GuildQueueService;
   private playerService: PlayerService;
-  private spotifyApi: SpotifyApi;
-  private youTubeIClient: YouTubeIClient;
-  private messageComponents;
+  private mediaService: MediaService;
 
   slashCommandConfig;
   commandNames = ["play"];
@@ -45,6 +37,8 @@ export default class PlayCommand implements Command {
   ) {
     this.playerService = playerService;
     this.queueService = queueService;
+    this.mediaService = new MediaService();
+
     this.slashCommandConfig = new SlashCommandBuilder()
       .setName(this.commandNames[0])
       .setDescription("Plays a song")
@@ -54,33 +48,6 @@ export default class PlayCommand implements Command {
           .setDescription("Song name or link")
           .setRequired(true)
       );
-
-    const nextSongButton = new ButtonBuilder()
-      .setCustomId(PlayCommandMessageComponentID.NEXT_SONG)
-      .setLabel("  ⏭️  ")
-      .setStyle(ButtonStyle.Primary);
-
-    const pauseSongButton = new ButtonBuilder()
-      .setCustomId(PlayCommandMessageComponentID.TOGGLE_PAUSE)
-      .setLabel("  ⏸  ")
-      .setStyle(ButtonStyle.Primary);
-
-    const prevSongButton = new ButtonBuilder()
-      .setCustomId(PlayCommandMessageComponentID.PREV_SONG)
-      .setLabel("  ⏮️  ")
-      .setStyle(ButtonStyle.Primary);
-
-    this.messageComponents = new ActionRowBuilder().addComponents(
-      prevSongButton,
-      pauseSongButton,
-      nextSongButton
-    );
-
-    this.spotifyApi = SpotifyApi.withClientCredentials(
-      process.env.SPOTIFY_CLIENT_ID,
-      process.env.SPOTIFY_CLIENT_SECRET
-    );
-    this.youTubeIClient = new YouTubeIClient();
   }
 
   async execute(interaction: ChatInputCommandInteraction) {
@@ -89,7 +56,9 @@ export default class PlayCommand implements Command {
 
       const inputValue = String(interaction.options.data[0].value);
 
-      const songs: SongInfo[] = await this.fetchRequestedMedia(inputValue);
+      const songs: SongInfo[] = await this.mediaService.fetchRequestedMedia(
+        inputValue
+      );
 
       const voiceChannel: VoiceBasedChannel =
         await this.getVoiceChannelFromInteraction(interaction);
@@ -101,22 +70,7 @@ export default class PlayCommand implements Command {
       });
 
       const serverQueue = this.queueService.getGuildQueue(interaction.guildId);
-
-      const messageComponents = new ActionRowBuilder().addComponents([
-        new ButtonBuilder()
-          .setCustomId(PlayCommandMessageComponentID.PREV_SONG)
-          .setLabel("  ⏪  ")
-          .setDisabled(serverQueue.songPosition === 0)
-          .setStyle(ButtonStyle.Primary),
-        new ButtonBuilder()
-          .setCustomId(PlayCommandMessageComponentID.TOGGLE_PAUSE)
-          .setLabel("  ▶  ")
-          .setStyle(ButtonStyle.Primary),
-        new ButtonBuilder()
-          .setCustomId(PlayCommandMessageComponentID.NEXT_SONG)
-          .setLabel("  ⏭️  ")
-          .setStyle(ButtonStyle.Primary),
-      ]);
+      const messageComponents = this.createMessageComponents(serverQueue);
 
       await interaction.editReply({
         embeds: [embed],
@@ -155,43 +109,46 @@ export default class PlayCommand implements Command {
           const nextSongembed = await this.playerService.nextSong(
             interaction.guildId
           );
+          const nextSongMessageComponents =
+            this.createMessageComponents(serverQueue);
+
           await interaction.update({
             embeds: [nextSongembed],
-            components: [this.messageComponents as any],
+            components: [nextSongMessageComponents as any],
           });
           break;
+
         case PlayCommandMessageComponentID.TOGGLE_PAUSE:
-          if (serverQueue.isPlaying == true) {
+          if (serverQueue.isPlaying === true) {
             await this.playerService.pauseSong({
               guildId: interaction.guildId,
             });
-
-            const messageComponents = this.createMessageComponents(serverQueue);
-
-            await interaction.update({
-              components: [messageComponents as any],
-            });
-            break;
           } else {
             await this.playerService.resumeSong(interaction.guildId);
-            const messageComponents = this.createMessageComponents(serverQueue);
-
-            await interaction.update({
-              components: [messageComponents as any],
-            });
-            break;
           }
+
+          const togglePauseMessageComponents =
+            this.createMessageComponents(serverQueue);
+
+          await interaction.update({
+            components: [togglePauseMessageComponents as any],
+          });
+          break;
+
         case PlayCommandMessageComponentID.PREV_SONG:
           const previousSongEmbed = await this.playerService.previousSong(
             interaction.guildId
           );
 
-          const messageComponents = this.createMessageComponents(serverQueue);
+          const previousSongMessageComponents =
+            this.createMessageComponents(serverQueue);
+
           await interaction.update({
             embeds: [previousSongEmbed],
-            components: [messageComponents as any],
+            components: [previousSongMessageComponents as any],
           });
           break;
+
         default:
           break;
       }
@@ -219,161 +176,6 @@ export default class PlayCommand implements Command {
         .setLabel("  ⏭️  ")
         .setStyle(ButtonStyle.Primary),
     ]);
-  }
-
-  async fetchYoutubeMedia(uri: string) {
-    if (uri.includes("list")) {
-      const urlParams = new URLSearchParams(uri);
-      const playlistID = urlParams.get("list");
-
-      const client = new YouTubeIClient();
-      const results = await client.getPlaylist(playlistID);
-      let videoItems: VideoCompact[];
-
-      if ("items" in results.videos) {
-        videoItems = results.videos.items;
-      } else {
-        videoItems = results.videos;
-      }
-
-      const songs: SongInfo[] = videoItems.map((element) => {
-        const songURL = new URL("https://www.youtube.com/watch");
-        songURL.searchParams.append("v", element.id);
-
-        return {
-          url: songURL.toString(),
-          title: element.title,
-          thumbnail: element.thumbnails[element.thumbnails.length - 1].url,
-          query: songURL.toString(),
-          duration: element.duration,
-          description: "From playlist",
-          uploader: "Uploader",
-        };
-      });
-
-      return songs;
-    } else {
-      const requestedMedia = await ytdl.getInfo(uri);
-
-      if (requestedMedia) {
-        const { videoDetails } = requestedMedia;
-        const { thumbnails } = videoDetails;
-        const thumbnail =
-          videoDetails.thumbnails[thumbnails.length - 1].url || null;
-
-        const song = {
-          url: videoDetails.video_url,
-          title: videoDetails.title,
-          thumbnail: thumbnail || " ",
-          description: videoDetails.title,
-          duration: Number(videoDetails.lengthSeconds),
-          query: uri,
-          uploader: videoDetails.author.name,
-        };
-
-        return [song];
-      }
-    }
-  }
-
-  async fetchSpotifyMedia(uri: string): Promise<SongInfo[]> {
-    const type = uri.split("/")[3];
-    const id = uri.split("/")[4];
-
-    if (type === "track") {
-      const result = await this.spotifyApi.tracks.get(id);
-      const { name, artists } = result;
-
-      const allArtists = artists
-        .flatMap((artist: Artist) => artist.name)
-        .join(" ");
-      const songName = `${allArtists} - ${name}`;
-
-      const song = {
-        url: result.uri,
-        duration: result.duration_ms / 1000,
-        title: songName,
-        thumbnail: " ",
-        description: "Spotify description",
-        query: songName,
-        uploader: "Spotify uploader",
-      };
-
-      return [song];
-    } else if (type === "album") {
-      const spotifyAlbum = await this.spotifyApi.albums.get(id);
-
-      const allArtists: string = spotifyAlbum.artists
-        .map((artist: Artist) => artist.name)
-        .join(" ");
-
-      const { items } = spotifyAlbum.tracks;
-      const songs: SongInfo[] = items.map((item) => {
-        return {
-          url: item.uri,
-          title: `${allArtists} ${item.name}`,
-          duration: item.duration_ms * 1000,
-          thumbnail: spotifyAlbum.images[0].url,
-          description: "Spotify description",
-          query: `${allArtists} ${item.name}`,
-          uploader: spotifyAlbum.label,
-        };
-      });
-
-      return songs;
-    } else if (type === "playlist") {
-      const playlist: any = await this.spotifyApi.playlists.getPlaylistItems(
-        id
-      );
-      const { items } = playlist.tracks;
-      const songs: SongInfo[] = items.map((item) => {
-        const { track } = item;
-        const artists = track.artists.map((artist) => artist.name).join(" ");
-
-        return {
-          url: track.uri,
-          title: `${artists} ${track.name}`,
-          duration: track.duration_ms / 1000,
-          thumbnail:
-            "https://storage.googleapis.com/pr-newsroom-wp/1/2018/11/folder_920_201707260845-1.png",
-          description: "Spotify description",
-          query: `${artists} ${track.name}`,
-          uploader: "Spotify uploader",
-        };
-      });
-
-      return songs;
-    }
-  }
-
-  public async fetchRequestedMedia(query: string): Promise<SongInfo[]> {
-    if (query.includes("youtu.be") || query.includes("youtube")) {
-      return await this.fetchYoutubeMedia(query);
-    } else if (query.includes("https://open.spotify.com")) {
-      return this.fetchSpotifyMedia(query);
-    } else {
-      const searchResults = await this.youTubeIClient.search(query);
-      const { items } = searchResults;
-
-      const topResult = items[0] as VideoCompact;
-      const { title, thumbnails, description, channel } =
-        items[0] as VideoCompact;
-
-      const songURL = new URL("https://www.youtube.com/watch");
-      songURL.searchParams.append("v", topResult.id);
-
-      const song: SongInfo = {
-        url: songURL.toString(),
-        title: title,
-        duration: topResult.duration,
-        thumbnail: thumbnails[topResult.thumbnails.length - 1].url,
-        description: description,
-        query: query,
-        uploader: channel.name,
-      };
-
-      return [song];
-    }
   }
 
   async getVoiceChannelFromInteraction(
